Cput

+++ src/nifti/niftilib/nifti1_io.c
@@ -2871,7 +2871,7 @@ int is_nifti_file( const char *hname )
 
    ii = znzread( &nhdr , 1 , sizeof(nhdr) , fp ) ;
    znzclose( fp ) ;
-   if( ii < sizeof(nhdr) )               return -1 ;  /* bad read? */
+   if( ii < (int) sizeof(nhdr) )               return -1 ;  /* bad read? */
 
    /* check for NIFTI-ness */
 
@@ -3245,7 +3245,7 @@ nifti_image* nifti_convert_nhdr2nim(struct nifti_1_header nhdr,
 
    if( is_onefile ){
      ioff = (int)nhdr.vox_offset ;
-     if( ioff < sizeof(nhdr) ) ioff = sizeof(nhdr) ;
+     if( ioff < (int) sizeof(nhdr) ) ioff = (int) sizeof(nhdr) ;
    } else {
      ioff = (int)nhdr.vox_offset ;
    }
